gkmamn/requirements.txt
flask
flask-fantasy